// src/components/Table/types.ts
export type SortOrder = 'asc' | 'desc';

export type Row = Record<string, string> & {
	_id?: string | number;
	id: string | number;
};

// Manage Columns
export type TComponentType =
	| 'link'
	| 'date'
	| 'dateTime'
	| 'duration'
	| 'status'
	| 'time';
// ...
export interface IColumn {
	type?: TComponentType;
	key: string;
	options?: Record<string, string>;
	field?: string;
	isVisible?: boolean;
	label: string;
	sortable?: boolean;
	colSpan?: number | string;
	render: () => React.ReactNode;
	isMandatory?: boolean;
}

export type IColumns = IColumn[];

export interface ITablePagination {
	currentPage?: number;
	limit: number;
	totalCount?: number;
}

export interface TTableFooterProps<T> {
	pagination: ITablePagination;
	loading?: boolean;
	changePagination: React.Dispatch<React.SetStateAction<number>>;
	onPageSizeChange?: () => void;
	rowsPerPageOptions?: number[];
	onPageChange?: (page: number) => {};
}

export interface IFetchTableDataArgs {
	page: number;
	limit?: number;
	sort?: { [x: string]: 1 | -1 };
}

export interface IProps<T> {
	columns: IColumns;
	fetchRowDetailData?: ({ rowId }: { rowId: string }) => Promise<any>;
	fetchTableData?: ({ page, sort }: IFetchTableDataArgs) => Promise<any>;
	loading: boolean;
	pagination: ITablePagination;
	renderRow?: () => React.ReactNode;
	renderRowDetail?: ({ columns, row }) => React.ReactNode;
	renderAttachment?: JSX.Element;
	linkComponent?: JSX.Element;
	rows: T[];
	withCollapsibleRows?: boolean;
	columnWidth?: string | number;
}

export type ITableHeaderProps<T> = Pick<IProps<T>, 'columns'> & {
	handleSort: (columnKey: IColumn['key']) => void;
	orderBy: IColumn['key'];
	sortOrder: SortOrder;
	singleSort?: boolean;
	rowDetail?: () => React.ReactNode;
	selectable: boolean;
	rowActions: () => React.ReactNode;
	handleSelectAll?: () => void;
	selectAllCheckbox?: boolean;
};